background image for NavBar method was changed
